Fix bug in place_piece_internal that sometimes duplicated pieces
Pieces are supposed to be sorted in order. If the piece was supposed to go in index 0, it would copy the old index 0 to index 1, but never write the new value to index 0. This effectively duplicated the old value of index 0.

Also add some convenience logic to fractogram_slicer that places slices
in the correct directory, not just the current directory.

Updated slice files will be committed separately.