Fix alpha
Before the behavior seemed to be "compute the color without alpha, and then mix that color with the background color from the HTML page". And so, e.g. transparent red ([1, 0, 0, 0]) would display as "pure red mixed 50/50 with whatever HTML element is behind it". With this actually displays as transparent, letting through 100% of the HTML behind it and 0% red, as I intended.